Push final

+++ mysite/datos_artetronica/models.py
@@ -368,9 +368,9 @@ ESTADO3 = (
 			)
 
 class Carro_de_compras(models.Model):	     
-	     id_usuario=models.CharField(max_length=30,blank=True)
-	     cantidad=models.FloatField(default=0,blank=True,null=True)
-	     producto=models.ForeignKey('Productos',blank=True,null=True)
+	     id_usuario=models.CharField(max_length=30,blank=True)	     
+	     cantidad=models.DecimalField(max_digits=4,decimal_places=0,default=0,blank=True,null= True)	     
+	     producto=models.ForeignKey('Productos',blank=True,null=True) 
 	     especificacion = models.TextField(blank=True,null=True)	     
 	     estado_prod=models.CharField(max_length=30,blank=True,choices=ESTADO3,default="EN_EXISTENCIA")
 	     fecha_ingreso = models.DateField(default=datetime.now,editable = False)
